refactor: methods structure in globals/api
