+++ src/globals/api/index.ts
@@ -4,7 +4,8 @@ export * from './methods/catalog-details'
 export * from './methods/filters'
 export * from './methods/geo'
 export * from './methods/building-progress'
-
+export * from './methods/complex'
+export * from './methods/complex-blocks'
 export * from './methods/sendCallRequest'
 
 // Mocked
