+++ src/globals/api/methods/catalog/index.ts
@@ -1,3 +1,2 @@
 export * from './getCatalogMenu'
-export * from './getComplexes'
 export * from './getCategory'
