+++ src/globals/api/methods/complex-blocks/getExclusiveOffers.ts
