+++ src/globals/api/methods/complex-blocks/index.ts
@@ -1,0 +1,5 @@
+export * from './getBonuses'
+export * from './getDocumentation'
+export * from './getExclusiveOffers'
+export * from './getLayouts'
+export * from './getYears'
