+++ src/globals/api/methods/complex/getComplexDetailed.ts
