+++ src/globals/api/methods/complex/getComplexHouseObjects.ts
@@ -1,6 +1,6 @@
 import { apiCall, APIRequest, APIResponse } from '@/globals/api/apiCall'
-import { CategoryRequestWithFilters, CategoryRequestWithPagination } from '@/types/catalog/Category'
 import { ComplexHouse, ComplexHouseObject } from '@/types/complex/ComplexHouse'
+import { GeneralRequestParams } from '@/types/RequestProps'
 import { Pagination } from '@/types/Pagination'
 
 type Response = APIResponse<
@@ -14,8 +14,7 @@ type Request = APIRequest<
     houseNumber: string
     orderBy?: 'asc' | 'desc'
     orderByWord?: 'flat_number' | string
-  } & CategoryRequestWithFilters &
-    CategoryRequestWithPagination
+  } & GeneralRequestParams
 >
 
 type Props = Omit<Request, 'houseNumber'>
