+++ src/globals/api/methods/complex/index.ts
@@ -1,0 +1,3 @@
+export * from './getComplexDetailed'
+export * from './getComplexes'
+export * from './getComplexHouseObjects'
